Add Base and priority calculators
To help give us insight into what the api wrappers will need to look
like.

I initially did the api wrapper (next commit) first but starting with
the calculaters makes it easier for us to know what we will be passing
into the wrapper - a package (I was wrong the first time).