Add basic USPS Ground Advantage calculator
This calculator will replace the old FirstClassMailParcel one that
existed in solidus_active_shipping. That delivery option was renamed and
merged into USPS Ground Advantage in 2023.

We'll likely still need to do a little bit more work ensuring the weight
limit is accurate for the new method and need some separate work to
ensure we're sending all of the required information along to USPS in
our API calls. (e.g.: Mail class is defined here but not used yet.)