Fill out remaining search data
After testing our old ActiveShipping implementation, it looks like we
just send 0 for all of the dimensions for all packages. So we're keeping
that logic here for now.

Because we have 0 for our dimensions, all packages are considered
"non-standard" and not "machinable." So that becomes the default
processing category for everything. (If we want to support flat rate
boxes, we'll need to do some tweaks to this logic I think.)

Our destination entry facility can just be "NONE" for everything. I
can't find any reference to this in the old API and it seems the most
reasonable option for us.

Finaly, the USPS API is faily particular about the values we send. Nil
is not allowed for any of the optional values, so we just need to
exclude them instead. Also, weight needs to be a number or integer
because strings aren't allowed.