Add mailing_date to search data
I had hoped that including this optional field would result in USPS only
returning rates that would be applicable for the mailing date.
Unfortunately, at least in their test environment that doesn't appear to
be the case.

Either way, its likely beneficial to include the value. We're using
"tomorrow" as a safe bet for when a package could concievably be
shipped.