Hardcode rate indicator to "SP" (single piece)
Using "LE" for first-class international caused an error in the API, but
"SP" works so we'll just use that for everything. Which allows us to
simplify some of our code a little bit.

Also updated the international calculator to only work for non-US
addresses and increased the weight limit based on what was used in the
old solidus_active_shipping gem.