Refactor #compute_package to the base calculator
We're performing the same operation for every calculator, just changing
the client and search data based on whether or not we need international
or domestic rates.

This moves all of the logic to the base and has each calculator define a
geo_group to specify whether it needs the international or domestic
pricing.

Also updates our calculators to return the amount, instead of the full
API request response which isn't useful for calculations up the line in
Solidus.